requirements.txt
selenium==4.27.1
webdriver-manager==4.0.2
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2024.1
